+++ Sources/CollisionObserver.swift
@@ -9,12 +9,6 @@
 /// Protocol to be implemented by objects that want to be notified of all
 /// collisions in a physics World
 public protocol CollisionObserver: class {
-    /// Called by the World to notify of a body collision
-    ///
-    /// - Parameter info: The information for the collision
-    @available(*, deprecated, message: "Use bodiesDidCollide(_:[BodyCollisionInformation]) instead.")
-    func bodiesDidCollide(_ info: BodyCollisionInformation)
-    
     /// Called by the World to notify of a complete set of body collisions in a
     /// single world update.
     ///