+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Adding `World.relaxWorld()` and `World.relaxBodies()` that allows relaxing all/a set of bodies into a more proper rest shape that will not 'pop' into their rest shape when starting to simulate. Can be called at any time during simulaton.
 - Making `BodyComponent` a protocol instead of a class.
+- Removing deprecated `CollisionObserver.bodiesDidCollide(_ info: BodyCollisionInformation)` protocol method.
 
 ## Stable Releases
 ---